feat: release on specific labels
Only generate a release if our semver related labels (`breaking`, `enhancement`, `fix`) or the `release` label are present on a merged pull request.

Changed from push (merge) on main branch to release generation happening when a pull_request is merged to main branch.

This gives us access to the pull requests labels without having to make API cals.

Currently we'd still need to label a pull request with `release` if it is a dependabot or manual pull request related to a CVE or security fix.

- [x] update CONTRIBUTING.md with new release information
- [x] manually add `vuln` and `release` labels to repository

Signed-off-by: jmeridth <jmeridth@gmail.com>