feat: Enable mirroring non-forked repos
This is a pretty simplistic change to add the ability to mirror
repos which are not actually forks of upstream projects. This is
useful for demos and testing but it also can be helpful if the
public organization actually *is* the upstream, because the repo
contains a project we own, so we're the top of the fork network.

Addresses #251